--- src/main/java/jp/xet/spring/aws/autoconfigure/AwsS3ClientProperties.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package jp.xet.spring.aws.autoconfigure;
-
-import lombok.Data;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- * TODO miyamoto.daisuke.
- *
- * @author miyamoto.daisuke
- * @since #version#
- */
-@Data
-@ConfigurationProperties(value = "aws.s3", ignoreInvalidFields = true)
-class AwsS3ClientProperties {
-	
-	private Boolean pathStyleAccessEnabled;
-	
-	private Boolean chunkedEncodingDisabled;
-	
-	private Boolean accelerateModeEnabled;
-	
-	private Boolean payloadSigningEnabled;
-	
-	private Boolean dualstackEnabled;
-	
-	private Boolean forceGlobalBucketAccessEnabled;
-	
-}
