--- src/main/java/jp/xet/spring/aws/autoconfigure/MutableEndpointConfiguration.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package jp.xet.spring.aws.autoconfigure;
-
-import lombok.Data;
-
-import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
-
-/**
- * @see <a href="https://github.com/spring-projects/spring-boot/issues/8762">spring-boot#8762</a>
- */
-@Data
-class MutableEndpointConfiguration {
-	
-	private String serviceEndpoint;
-	
-	private String signingRegion;
-	
-	
-	EndpointConfiguration toEndpointConfiguration() {
-		if (serviceEndpoint != null) {
-			return new EndpointConfiguration(serviceEndpoint, signingRegion);
-		}
-		return null;
-	}
-}
